mediatek: Check whether a previous strategy is active using a delay for output-switching cases
In the mediatek stack the flow for changing between speaker and headset in the FM
application operates via forcing the media strategy to output to the speaker.

The default behavior for the policy on force-use cases is to simply go reassign
devices according to whatever strategy is active. Unfortunately, for the case of
media, and in the specific case of forced routing, the check for streams on the
current strategy simply ignored whether the stream was playing in the recent past.

We previously solved this by ignoring the check for the recent past but this
causes the media stream to be active at all times. This patch changes the behavior
slightly and now we check whether the stream was actually running when evaluating
the creation of new devices for media strategies.

Change-Id: I6b2ca9ce3c3f905c6c29cc31026d6643b0de434e